+++ Source/storm/storm_net.hpp
@@ -87,7 +87,7 @@ bool SNetJoinGame(char *gameName, char *gamePassword, int *playerid);
  */
 bool SNetLeaveGame(int type);
 
-bool SNetReceiveMessage(int *senderplayerid, void **data, uint32_t *databytes);
+bool SNetReceiveMessage(uint8_t *senderplayerid, void **data, uint32_t *databytes);
 bool SNetReceiveTurns(int arraysize, char **arraydata, size_t *arraydatabytes, uint32_t *arrayplayerstatus);
 
 typedef void (*SEVTHANDLER)(struct _SNETEVENT *);
